requirements.txt
annotated-types==0.7.0
appdirs==1.4.4
asttokens==3.0.1
attrs==25.4.0
build==1.3.0
cachey==0.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
configobj==5.0.9
DateTime==5.5
decorator==5.2.1
dllist==2.0.0
docstring_parser==0.17.0
einops
executing==2.2.1
filelock==3.19.1
flatbuffers==25.9.23
flexcache==0.3
flexparser==0.4
future==1.0.0
greenlet==3.2.4
HeapDict==1.0.1
hsluv==5.0.4
humanfriendly==10.0
idna==3.11
imagecodecs==2025.8.2
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
in-n-out==0.2.1
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lark==1.3.1
lazy_loader==0.4
linecache2==1.0.0
lxml==6.0.2
MarkupSafe==2.1.5
mpmath==1.3.0
msgpack==1.1.2
ndindex==1.10.0
nibabel==5.3.3
ninja==1.13.0
numexpr==2.14.1
numpy==2.2.6
onnx==1.19.1
onnx-simplifier==0.4.36
onnxruntime==1.23.2
opencv-contrib-python==4.12.0.88
packaging==25.0
pillow==12.0.0
Pint==0.25.2
platformdirs==4.5.0
pooch==1.8.2
portalocker==3.2.0
protobuf==6.33.0
psutil==7.1.2
pure_eval==0.2.3
puremagic==1.30
py-cpuinfo==9.0.0
pybind11==3.0.1
pyconify==0.2.1
pydantic==2.12.5
pydantic-compat==0.1.2
pydantic_core==2.41.5
PyOpenGL==3.1.10
pyparsing==3.2.5
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pywin32==311
PyYAML==6.0.3
referencing==0.37.0
reportlab==4.4.10
requests==2.32.5
roman-numerals-py==3.1.0
rpds-py==0.29.0
scipy==1.16.3
setproctitle==1.3.7
shellingham==1.5.4
six==1.17.0
stack-data==0.6.3
termcolor==3.2.0
tifffile==2025.10.16
timm==0.4.12
tomli_w==1.2.0
torch==2.9.0
torchvision==0.24.0
tqdm==4.67.1
traceback2==1.4.0
typer==0.9.0
typing_extensions>=4.14.1
tzdata==2025.1
uc-micro-py==1.0.3
ujson==5.10.0
unittest-xml-reporting==3.2.0
urllib3==2.3.0
vtk==9.5.2
wcwidth>=0.2.13
wrapt==1.17.2
yacs==0.1.8
zipp==3.21.0
